Fix t:datefield for bootstrap-5 migration
In sirius.kernel.commons.AdvancedDateParser#ensureValidYear we only allow values between 1900 - 2100. So we also can do this in the Frontend via min and max values at the date input.
Also the oninvalid event must reset the single-click.pending-js on the save button, because otherwise (after an invalid input) we can never press save-button again,
The value now will nomore be parsed twice, so we have to fix a breaking change....

HINT: copy-pasting into a datefield is with the date-input no more available (in contrast to the before-bs5-version) and not easy to fix. Maybe a small hack like onpaste()-event could be a solution, but input-sanitizing would be the next problem then...

BREAKING: We have to use @toMachineString(dateFieldValue) for all values in t:datefields now.

- fixes: SIRI-938