Fixes authentication errors against SMTP servers which require early passwords.
We already provide the password via an authenticator implementation (sirius.web.mails.SendMailTask.MailAuthenticator). However SMTPTransport#sendMessage does evaluate the password set this way only after the initial connection.

The transport seems to normally challenge the client for a password which is then correctly set already. However for some SMTP servers (e.g. SMTP relays with strict rules) which require early authentication, this is not enough.

Setting the password at the transport will in turn set the password on a URLName object which is getting evaluated early.

Fixes: SE-13522