Only detect intrinsic NLS get call for single parameter calls
The previous logic also optimized NLS get calls with multiple parameters (there is a variant with an int for singular/plural and one with a string for a forced language) to the NLS get call with the single string parameter for the key.
This meant that the other two variants were not callable in Noodle scripts and Tagliatelle templates.

Fixes: OX-10611