Passes along previous parameters when parsing the next one
Using the previous parameters, we can - in case of a lambda - better infer generics if
these are pre-determined by previous parameters.

If we have a signature like <E> void foo(Class<E> type, Consumer<E> consumer) which
is invoked via foo(String.class, | s | ...) then we know, as w're parsing the lambda, that E
is a String (as this is defined by the first parameter. This is now propagated along by this
change.