Delay logging of error message when invalid JSON content is fetched
For some calls it may not be easy to determine if it contains a JSON body or we want to try to access it anyway.
With this change the caller can decide whether to log the thrown exception or not.

The previous solution of commit 0d19e75c1a56a33a22584b05f0d169c11de10ce5 was incomplete in this regard as Json#parseObject always logs an error message to the main logger.

Fixes: OX-10940