Adds link and title properties 🔧
When using `<t:iconInlineInfo>` within `<t:datacard>`, users may want to know more about the information displayed, depending on how obvious the icon is. The new `title` property allows to set such transformation, which will be displayed as BS tooltip and rendered as ARIA label.

In order to make this work, the tooltip technically needs to become a link, thus overriding the card's own link. It is thus convenient to also pass the link property to allow seamless navigation.