Uses uri resolving of the UserAgentCallback in the third attempt
The newly implemented option of resolving images very efficiently does not work completely reliably in a few cases. For example, if we have URLs that are specified with http but automatically redirect to https. Although this happens very rarely, it should still work. The old, more expensive method via the UserAgentCallback handles this very well. As we only need this in a few exceptions, we only use the fallback if it has not worked twice with the new method.

Fixes: DH-961