Align logic of get function in QueryString with WebContext
This is a BREAKING api change of the helper so they behave the same when a parameter is provided multiple times in a request (as in has multiple values).

It is done to be less confusing by having two similar named functions with different behaviour and so the helper could be used in the WebContext itself.

Fixes: SIRI-963