fix t:radio - correct classes and according selector.
input-required does not exist anymore in BS5, from-text should be form-text
Many radio-buttons with a required-* are not useful.
With this approach you can add a required-* to a div with a t:radio inside as well, but the * will only occur on the heading label, not at the selectable values.

- fixes: SE-13684