Prevent closing when click is dragged from within overlay to background
The previous behaviour is a result of this caveat of the click event:

If the button is pressed on one element and the pointer is moved outside the element before the button is released, the event is fired on the most specific ancestor element that contained both elements.

So if a click started within the overlay content and was dragged outside of it to the background (which can happen by accident when content with drag support is displayed inside the overlay) the overlay would suddenly close.

We now use the mousdown event so it only triggers when the click is started outside of the overlay content.

Tested on Desktop Chrome, Android Chrome & IE 11.

Fixes: OX-11077