Allow to send POST data as urlencoded instead of multipart form data
Netty seems to replace some characters like commas, colons, semicolons, ... with spaces when reading the multipart form data in `HttpPostMultipartRequestDecoder::getContentDispositionAttribute`.
This may be problematic when trying to match the parameter names to data.

To circumvent this in a non breaking fashion we add a new optional parameter that converts the FormData into an URLSearchParams object.
This way data is sent with content type `application/x-www-form-urlencoded` instead of `multipart/form-data`.

Fixes: OX-10757