Passes valid XHTML to the PDF renderer 🚔
The PDF renderer requires XHTML as input, and we used to rely on the invoker to provide valid content. However, luckily, the world has moved on from XHTML. In order to relax the requirements and facilitate reuse of HTML templates, this commit introduces an additional processing step that removes problematic elements and ensures XHTML automatically.

OX-10783