Reorder design system rules to enable more flexible overwriting
CSS rules of the same specificity (most of our helpers only use a single class, so have the specificity 0,1,0) are applied in the order they are loaded/defined.
This means that we should place more specific helpers below more board components to increase the flexibility.
Doing so reduces the need for !important markings on rules, which can get quite out of hand.

Examples of what the proposed changes allow to do:

1) `sci-button-primary sci-p-1` - allows to reduce the default padding of our button class

2) `sci-d-flex sci-d-none` - Allows to hide an element and show it as soon as the `sci-d-none` is removed (the JS code does not need to know what proper display class has to be added)

Fixes: OX-11120