Extract sirius text sizes into CSS variables
This will allow us to transform them into using pixels as a unit instead of rem.
Reason:
We use the CSS design system for our integrations into other websites, where we don't have control over the base font size. Because of this a bunch of customers have problems with small text / elements / margins, as they specify a base font size of 10 px instead of 16 px.
To mitigate this we want to switch to pixels and allow customers to easily alter the used sizes.
This can be realized most flexible by using CSS variables.

Fixes: OX-11040