Switch defined font sizes from rem to px
This replaces the sci-enable-font-size marker class that is obsolete with the new approach.

Reason:
We use the CSS design system for our integrations into other websites, where we don't have control over the base font size. Because of this a bunch of customers have problems with small text / elements / margins, as they specify a base font size of 10 px instead of 16 px.
To mitigate this we switch to pixels and allow customers to easily alter the used sizes.

Fixes: OX-11040