Base text size on base unit and remove display size scaling
This was taken from the enable-text-sizes mechanism but was not actually what was happening without the marker class.
If a user of the design system wants similar behaviour he could re-introduce these few media queries quite easily.

Also the text sizes are now based on the base unit so nearly the whole UI can be scaled by overriding the base unit CSS variable.

Fixes: OX-11040