Allow to provide a schema download next to the sample response of an API
This way a JSON-schema or XSD can be provided to further document an API response.

We use the @Schema sub annotation of the swagger ApiResponse annotation for this.
A "ref" has to be filled with the downloadable asset URI of the schema.
The "name" field can be used to provide an optimized download name for the schema file.

Fixes: OX-11872