Introduce new way of coloring CSS SVG icons
Defined via mask-image and background-color results in the icons automatically following the current foreground color.
This makes them much easier to use, especially when embedding them in elements with a color change on hover, like our button elements.
Additionally we don't have to define a new class when we need an icon in another color, they are now automatically compatible with all text color classes.

Fixes: OX-11941