Correctly resolves the X509 certificate 🔍
Previously, we expected the X509Certificate to be the first child of the X509Data structure. For Google SAML, this assumption is wrong as the X509Certificate is the second entry only, leading to an illegal cast. Therefore, we now inspect all children and process the first X509Certificate child.

OX-12034