Allows controllers to handle exceptions custom.
Some controllers may not want to log exceptions to the system log or they want to enrich the exception before logging it, e.g. adding a generic error code so that later on, a more generic error message can be displayed to the user.

By default, the exception is simply handled via the common logic as was the case already. Because handled exceptions are ignored returned as is when calling Exceptions#handled, we don't have to differentiate between those and unhandled exceptions in the BasicController.

Fixes: SE-14652