Fixes losing the scope when retrying to send a mail.
This is normally not an issue as the mail content is already rendered in the user language and co. and the mail settings are stored in the SendMailTask. However, the OAuth access token may need to be retrieved on a retry and because it depends on the current scope, we mustn't lose the scope!

Fixes: SIRI-1060