Optionally allows to request SAML auth for a specific user 👮‍♂️
Some (few) SAML servers allow to specify a user for whom to request SAML authentication while others reject this as erroneous request. This commit, therefore, allows to optionally include the user in the request, but a potential invoker will need to make sure that this is done for compatible providers only.

MIO-6450