Guesses API format when RAW is given 🤔
An API route annotated with `@PublicService` can have one of three formats, JSON, XML and RAW. For the former two, Sirius creates a `StructuredOutput` that can be easily populated. However, the price is that the HTTP response code is fixed to 200 (OK) and custom response headers are not possible.

When implementing a RESTful API, however, we need other response codes from the 2xx range as well, particularly 201 (Created) and 204 (No Content). Also, custom headers are sometimes required. A simple way to still achieve that, is to declare the format as RAW and write the JSON oneself. In that case, though, the raw format will show up in the API documentation.

Hence, this commit adds a facility to guess the effective return format from the documented API responses in case the programmer declared the output to be RAW. In case all documented API responses declare a JSON media type (except for examples without a media type), the effective format will be JSON. Likewise for XML. If no consistent format can be determined, the code will fall back to RAW.

Note that this affects the API documentation module only.

MIO-6426